changed volume data type to double

+++ src/delivery/domain/Category.java
@@ -3,7 +3,7 @@ package delivery.domain;
 public class Category {
     private int id;
     private String type;
-    private int volume;
+    private double volume;
 
     public Category(int id, String type, int volume) {
         this.id = id;
@@ -27,11 +27,11 @@ public class Category {
         this.type = type;
     }
 
-    public int getVolume() {
+    public double getVolume() {
         return volume;
     }
 
-    public void setVolume(int volume) {
+    public void setVolume(double volume) {
         this.volume = volume;
     }
 }
